fix public projects
