+++ src/app/app/virtual-lab/(home)/public-projects/[slug]/page.tsx
@@ -12,7 +12,7 @@ export default async function SinglePublicProjectPage({ params }: ParamProps) {
 
   return (
     <>
-      <PublicProjectSideNavigation />
+      {/* <PublicProjectSideNavigation /> */}
       <PublicProjectMain slug={slug} />
     </>
   );
